refactor proficiency into feature module
