+++ src/features/proficiency/state.js
@@ -1,0 +1,3 @@
+export const proficiencyState = {
+  proficiency: {},
+};
